refactor(db): rework the ULID migrations around derived timestamps and no-`ts_ingest` helpers
Change the base and hypertable migrations to derive `ts_ingest` from `id`, make JSON-schema validation optional, switch manifest and promotion-queue FKs to ULID-aware forms, delete the experimental `20250103120013` migration pair, add no-`ts_ingest` model and query variants, and adjust the reset/setup scripts plus worker crates to match the rewritten database contract.

+++ src/sinex-worker/src/worker.rs
@@ -65,7 +65,7 @@ impl Worker {
             &self.pool,
             self.processor.agent_name(),
             &self.worker_id,
-            self.processor.batch_size(),
+            self.processor.batch_size() as i64,
         )
         .await?;
 
@@ -79,7 +79,7 @@ impl Worker {
             count = count,
             "Claimed items for processing"
         );
-        self.metrics.items_claimed.inc_by(count as u64);
+        self.metrics.items_claimed.inc_by(count as f64);
 
         for item in items {
             let start = std::time::Instant::now();
